feat: PWM DAC on GP46, with the speaker and the OPL2 mixed into it
GP46 is all the audio this board has, and the PC speaker and the AdLib
both want it. So it stops being a square-wave output and becomes a DAC: a
~1.97 MHz carrier whose duty is rewritten at 44.1 kHz by DMA, with both
sources summed in software.

Two DMA channels ping-pong between two 256-sample buffers, chained to
each other and paced by a DMA fractional timer, so nothing is serviced
per sample -- the interrupt only records which buffer drained, and core 0
refills it from the main loop rather than on the frame tick, since a
block is 5.8 ms and a frame is 16.6.

The pacing fraction is worth a note: the timer is sys_clk * num/den with
both fields 16 bits, and 44100/504000000 reduces to 7/80000, which does
not fit. 1/11429 gives 44098.3 Hz, which is 0.004 % low. It is computed
from the live clock, so changing CPU_SPEED does not silently retune every
sound the machine makes.

This is an improvement to the speaker independently of the OPL. The old
path set the PWM *wrap* register to the PIT's channel-2 reload value,
which tied the audio resolution to the guest's chosen pitch: eleven bits
of duty at a typical reload of 1193, seven at a high-pitched 100. The
square wave is now synthesised at a fixed sample rate and the two are
unrelated.

Both emulated chips only record numbers now -- i8253 the divisor, port
61h the gate -- because both run in the bus interrupt with the 8086 in a
wait state, and reprogramming a peripheral there was never a good idea.

One trap, for the next person: audio takes DMA_IRQ_1 because the VGA
driver installs an *exclusive* handler on DMA_IRQ_0. Registering a shared
handler there first makes VGA's irq_set_exclusive_handler() panic, and
with PICO_PANIC_FUNCTION=0 a panic is a silent hang -- the symptom was
simply that the boot stopped after the audio line with nothing to say
why.

+++ src/chipset/adlib.h
@@ -60,10 +60,14 @@ uint8_t adlib_status(void);
 // ---------------------------------------------------------------------------
 
 /*
- * Render `nsamples` mono samples, signed 16-bit, into `out`.
+ * Render `nsamples` mono samples into `out`.
+ *
+ * 32-bit, because EMU8950_LINEAR selects OPL_calc_buffer_linear() and
+ * that is the width it produces -- and the mixer sums in 32 bits anyway,
+ * so taking it in the native width saves a copy and a clip.
  *
  * Returns false when there is nothing to render — no chip, or the guest
  * has never touched the card — so the caller can skip mixing entirely
  * and a machine that never plays a note costs nothing.
  */
-bool adlib_render(int16_t *out, uint32_t nsamples);
+bool adlib_render(int32_t *out, uint32_t nsamples);
